README: document impact of balance + quota groups
The combination of balance and quota groups as the number of references
to extents rises (either via snapshots, dedupe, or reflink copies) can
be extreme.

The file system performs quota group accounting for every extent relocated
and the result can be the appearance of stalled system as the
transaction commit can take a very long time.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>