sysconfig: balance period defaults to 'none' (fate#324529).
It used to be that we'd run balance periodically because otherwise
empty (or nearly empty) block groups wouldn't be cleaned up.  That hasn't
been the case for some years now, so we don't need to run it automatically.

Users who have workloads that swing wildly between metadata and data
intensive may want to enable this on their systems but most users
will not need it.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>