btrfs-defrag-plugin: add shell implementation
Requiring zypp-plugin-python for the python plugin means that we need to
pull in 150 MB of python dependencies.  It's present already on many
systems but on JeOS systems, it won't be by default.  We still want to
defrag the RPM database on those systems.

This commit adds the shell script that provides the same functionality.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>