btrfs-defrag-plugin: remove dependency on zypp-plugin-python (jsc#SLE-4130)
Requiring zypp-plugin-python means that we need to pull in 150 MB of
python dependencies.  It's present already on many systems but on
JeOS systems, it won't be by default.  We still want to defrag
the RPM database on those systems.

This commit replaces the python script with a shell script that provides
the same functionality and drops the python dependency entirely.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>