Prevent concurrent balance+defrag+scrub+trim (boo#1088010)
When the services are triggered at the same time the access to the
physical devices is the limiting factor which can cause a severe
performance impact on the system.

The statements of "After=" do not prevent this when the services are
triggered individually, for example by the corresponding timer units,
hence they would start in parallel. Specifying a more fine-grained timer
schedule with selected operation windows will still trigger the services
at the same time when they could not be triggered earlier, for example
when the system was suspended so we need to handle the exclusive access
within the script calls.

Related bug: https://bugzilla.suse.com/show_bug.cgi?id=1088010