Make balance, scrub, and trim mutually exclusive tasks
When scheduled simultaneously, balance, scrub, and trim can cause the
file system to stall due to balance generating large numbers of
transactions and trim taking the commit root semaphore across
potentially time consuming operatations, which in turn limits the when
transactions can be committed.

The original idea for this was from Oliver Kurz <okurz@suse.com>, who
wrote it for systemd unit files.

This patch adds a file-lock based protection to the tasks that will
serialize on that in case the timers start the scripts before another
task finishes. This works for systemd timers and for cron.

Pull-request: #75
Issue: #15
Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>