btrfsmaintenance-refresh-cron.sh: reset OnCalendar for generated drop-in files
Add an empty OnCalendar line to the generated drop-in conf file. This
fixes the linked issues where the default periods are used in addition
to any periods provided in the btrfsmaintenance config, which is not
what users expect.

See systemd.timer(5) why it needs to be reset.

Issue: #71
Issue: #91
Pull-request: #94
Author: David Wilkinson <davidmwilkinson@protonmail.com>
Signed-off-by: David Sterba <dsterba@suse.com>