btrfsmaintenance-refresh-cron.sh: try three times to check if the configuration file exists
When tuning the value BTRFS_BALANCE_PERIOD in
/etc/sysconfig/btrfsmaintenance, btrfsmaintenance-refresh.service will
be triggered. But sometimes it will fail because the script
/usr/share/btrfsmaintenance/btrfsmaintenance-refresh-cron.sh claims the
file /etc/sysconfig/btrfsmaintenance does not exist. Because there's a
timing window between the service is triggered and the configuration
file is checked.
"""
Setting up watches.
Watches established.

/etc/sysconfig/ OPEN
/etc/sysconfig/ CLOSE_NOWRITE,CLOSE
/etc/sysconfig/ OPEN
/etc/sysconfig/ ACCESS
/etc/sysconfig/ CLOSE_NOWRITE,CLOSE
/etc/sysconfig/ MOVE_SELF
/etc/sysconfig/ ATTRIB
/etc/sysconfig/ DELETE_SELF
"""
As a result it will be chaos after the service file finishes.

Reference: bsc#1235954