Randomize mount point order for 'auto'
As reported:

If one uses auto to detect filesystems to act one, it might be a good
idea to randomize them.  If things go wrong and the operation is stopped
too early, then things are left (for next time). Randomizing the order
gives a chance that next time the left out things come first and have a
better chance.

The suggested fix is to shuffle the paths.

Issue: #135
Signed-off-by: David Sterba <dsterba@suse.com>